refactor: restructure composable logic
