+++ src/runtime/routing/routing.ts
@@ -4,7 +4,7 @@ import { assign, isString } from '@intlify/shared'
 import type { Locale } from 'vue-i18n'
 import type { RouteLocationNamedRaw, RouteLocationPathRaw, RouteLocationRaw } from 'vue-router'
 import type { CompatRoute } from '../types'
-import type { ComposableContext } from '../utils'
+import type { ComposableContext } from '../composable-context'
 
 export type RouteLikeWithPath = RouteLocationPathRaw & { name?: string }
 export type RouteLikeWithName = RouteLocationNamedRaw & { path?: string }
